carmel 3.11 compose segfault fixed
git-svn-id: https://nlg0.isi.edu/svn/sbmt/trunk/graehl@2188 7f356e55-ce0a-0410-be72-bc6dfe96aa1c

+++ carmel/src/carmel.cc
@@ -647,7 +647,7 @@ main(int argc, char *argv[]){
     }
     istream *pairStream = NULL;
     bool train_cascade=long_opts["train-cascade"];
-    cascade_parameters cascade(train_cascade,long_opts["debug-cascade"]);
+    cascade_parameters cascade(train_cascade,(unsigned)long_opts["debug-cascade"]);
     if ( train_cascade )
         flags['t']=1;
     if ( flags['t'] )
